+++ changelog.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Next
-- Updated resource pack format to `34`.
-- Updated data pack format to `48`.
+- Updated resource pack format to `36`.
+- Updated data pack format to `50`.
 - Fixed special case checking not being applied to full version names leading to no result for certain versions.
 
 ## 1.3.15